+++ app/ssiForm/constants.ts
@@ -1,0 +1,17 @@
+export const days = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16-30', '31-60', '61-90'];
+
+export const symptoms = [
+  'purulentDischarge',
+  'localizedPain',
+  'localizedSwelling',
+  'fever',
+  'incisionOpened',
+  'dehiscence',
+  'abscess',
+  'microorganismsSeen',
+  'imagingEvidence',
+  'positiveCulture',
+  'bloodCultureSent',
+  'diagnosisSSI',
+  // 'anyOther'
+];
