+++ app/ssiForm/microbiology_data.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { FormData } from './page';
 import DropdownBox from '../components/DropdownBox';
