+++ app/ssiForm/page.tsx
@@ -6,6 +6,7 @@ import MicrobiologyData from './microbiology_data';
 import AntibioticPrescription from './antibiotic_prescription';
 import './style.css';
 import PostOp_Sheet from './postop_form';
+import { days, symptoms } from './constants';
 
 interface Antibiotic {
   abop_stage: 'prior' | 'pre_peri' | 'after';
@@ -15,23 +16,6 @@ interface Antibiotic {
   doses: number;
 }
 
-export const days = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16-30', '31-60', '61-90'];
-
-export const symptoms = [
-  'purulentDischarge',
-  'localizedPain',
-  'localizedSwelling',
-  'fever',
-  'incisionOpened',
-  'dehiscence',
-  'abscess',
-  'microorganismsSeen',
-  'imagingEvidence',
-  'positiveCulture',
-  'bloodCultureSent',
-  'diagnosisSSI',
-  // 'anyOther'
-];
 
 const symptomsDict: { [key: string]: { [key: string]: boolean } } = {};
